feat(deep-link): land hub/run on the full repo file list
Replace the single-file confirm sheet with a global host that resolves
the whole repo via resolveHFRepo and presents the existing DetailsView
quant list, so the user picks a file (per-file download/progress is
owned by ModelFileCard, unchanged). The HF deep link has no quant, so a
single-file landing was the wrong default. Host shows a spinner while
resolving and an inline error with Retry on failure; dismiss clears the
parked request.
